Could not get this working without a bower install
Heyya:

Just tried to get this working based on your instructions and ran into a lot of errors that seemed as if they were caused by missing dependencies.  After installing bower and running a bower install on the project I had everything that ```yarn storybook``` needed in order to run properly.